+++ R/fps.R
@@ -1,16 +1,28 @@
-.xrd_nnls <- function(xrd.lib, xrd.sample, force) {
+.xrd_nnls <- function(xrd.lib, xrd.sample, force, tth_fps) {
 
   if(missing(force)) {
     force <- c()
   }
 
-  #Do the initial NNLS
-  mat <- xrd.lib$xrd
+  #check for excluded tth values
+  ex_tth <- which(xrd.lib$tth < tth_fps[1] | xrd.lib$tth > tth_fps[2])
 
-  x <- nnls::nnls(as.matrix(mat), xrd.sample)
+  xrd_sub <- xrd.lib$xrd
+
+  smpl_sub <- xrd.sample
+
+  if (length(ex_tth) > 0) {
+
+    #Remove the rows that are outside of tth_fps range
+    xrd_sub <- xrd_sub[-ex_tth,]
+    smpl_sub <- xrd.sample[-ex_tth]
+
+  }
+
+  #Do the initial NNLS
+  x <- nnls::nnls(as.matrix(xrd_sub), smpl_sub)
   x <- x$x
-  names(x) <- names(data.frame(mat,
-                               check.names = FALSE))
+  names(x) <- names(xrd_sub)
 
   #If force isn't being used then just remove any phase = 0
 
@@ -26,7 +38,20 @@
 
 }
 
-.fullpat <- function (par, pure_patterns, sample_pattern, obj) {
+.fullpat <- function (par, pure_patterns, sample_pattern, obj,
+                      tth, tth_fps, weighting) {
+
+  #check for excluded tth values
+  ex_tth <- which(tth < tth_fps[1] | tth > tth_fps[2])
+
+  if (length(ex_tth) > 0) {
+
+    #Remove the rows that are outside of tth_fps range
+    pure_patterns <- pure_patterns[-ex_tth,]
+    sample_pattern <- sample_pattern[-ex_tth]
+    weighting <- weighting[-ex_tth]
+
+  }
 
   #if only 1 pattern is being fitted:
   if (length(par) == 1) {
@@ -36,18 +61,25 @@
 
     #objective functions
     if(obj == "Delta") {
-      d <- sum(abs(sample_pattern - s_mix))
+      d <- delta(measured = sample_pattern,
+                 fitted = s_mix,
+                 weighting = weighting)
     }
 
     if(obj == "R") {
-      d <- sqrt(sum((sample_pattern - s_mix)^2)/sum(sample_pattern^2))
+      d <- r(measured = sample_pattern,
+             fitted = s_mix,
+             weighting = weighting)
     }
 
     if(obj == "Rwp") {
-      d <-  sqrt(sum((1/sample_pattern) * ((sample_pattern - s_mix)^2)) / sum((1/sample_pattern) * (sample_pattern^2)))
+      d <-  rwp(measured = sample_pattern,
+                fitted = s_mix,
+                weighting = weighting)
     }
 
     return(d)
+
   }
 
   #if more than 1 pattern is being fitted
@@ -59,15 +91,21 @@
 
     #objective functions
     if(obj == "Delta") {
-      d <- sum(abs(sample_pattern - s_mix))
+      d <- delta(measured = sample_pattern,
+                 fitted = s_mix,
+                 weighting = weighting)
     }
 
     if(obj == "R") {
-      d <- sqrt(sum((sample_pattern - s_mix)^2)/sum(sample_pattern^2))
+      d <- r(measured = sample_pattern,
+             fitted = s_mix,
+             weighting = weighting)
     }
 
     if(obj == "Rwp") {
-      d <-  sqrt(sum((1/sample_pattern) * ((sample_pattern - s_mix)^2)) / sum((1/sample_pattern) * (sample_pattern^2)))
+      d <-  rwp(measured = sample_pattern,
+                fitted = s_mix,
+                weighting = weighting)
     }
 
     return(d)
@@ -83,7 +121,7 @@
 #' Applies full pattern summation (Chipera & Bish, 2002, 2013; Eberl, 2003) to an XRPD
 #' measurement to quantify phase concentrations. Requires a \code{powdRlib} library of
 #' reference patterns with reference intensity ratios in order to derive
-#' mineral concentrations.
+#' mineral concentrations. Details provided in Butler and Hillier (2021).
 #'
 #' @param lib A \code{powdRlib} object representing the reference library. Created using the
 #' \code{powdRlib} constructor function.
@@ -171,6 +209,9 @@
 #'
 #' }
 #' @references
+#' Butler, B. M., Hillier, S., 2021.powdR: An R package for quantitative mineralogy using full pattern
+#' summation of X-ray powder diffraction data. Comp. Geo. 147, 104662. doi:10.1016/j.cageo.2020.104662
+#'
 #' Chipera, S.J., Bish, D.L., 2013. Fitting Full X-Ray Diffraction Patterns for Quantitative Analysis:
 #' A Method for Readily Quantifying Crystalline and Disordered Phases. Adv. Mater. Phys. Chem. 03, 47-53.
 #' doi:10.4236/ampc.2013.31A007
@@ -194,7 +235,7 @@ fps <- function(lib, ...) {
 #' Applies full pattern summation (Chipera & Bish, 2002, 2013; Eberl, 2003) to an XRPD
 #' sample to quantify phase concentrations. Requires a \code{powdRlib} library of reference
 #' patterns with reference intensity ratios in order to derive mineral
-#' concentrations.
+#' concentrations. Details provided in Butler and Hillier (2021)
 #'
 #' @param lib A \code{powdRlib} object representing the reference library. Created using the
 #' \code{powdRlib} constructor function.
@@ -238,6 +279,12 @@ fps <- function(lib, ...) {
 #' @param remove_trace A single numeric value representing the limit for the concentration
 #' of trace phases to be retained, i.e. any mineral with an estimated concentration below
 #' \code{remove_trace} will be omitted. Default = 0.
+#' @param weighting an optional 2 column data frame specifying the 2theta values in the first
+#' column and a numeric weighting vector in the second column that specifies areas of the pattern
+#' to either emphasise (values > 1) or omit (values = 0) when minimising the objective function
+#' defined in the \code{obj} argument. Use this weighting parameter with caution. The default
+#' is simply a weighting vector where all values are 1, which hence has no effect on the computed
+#' objective function.
 #' @param ... other arguments
 #'
 #' @return a list with components:
@@ -322,6 +369,9 @@ fps <- function(lib, ...) {
 #'
 #' }
 #' @references
+#' Butler, B. M., Hillier, S., 2021.powdR: An R package for quantitative mineralogy using full pattern
+#' summation of X-ray powder diffraction data. Comp. Geo. 147, 104662. doi:10.1016/j.cageo.2020.104662
+#'
 #' Bish, D.L., Post, J.E., 1989. Modern powder diffraction. Mineralogical Society of America.
 #'
 #' Chipera, S.J., Bish, D.L., 2013. Fitting Full X-Ray Diffraction Patterns for Quantitative Analysis:
@@ -337,7 +387,7 @@ fps <- function(lib, ...) {
 #' @export
 fps.powdRlib <- function(lib, smpl, harmonise, solver, obj, refs, std, force, std_conc, normalise,
                 tth_align, align, manual_align, tth_fps, shift,
-                remove_trace, ...) {
+                remove_trace, weighting, ...) {
 
 #---------------------------------------------------
 #Conditions
@@ -593,6 +643,24 @@ fps.powdRlib <- function(lib, smpl, harmonise, solver, obj, refs, std, force, st
 
   }
 
+
+  #Check the weighting data
+  if (missing(weighting)) {
+
+    weighting <- data.frame("tth" = smpl[[1]],
+                            "counts" = 1)
+
+  }
+
+  if (!is.data.frame(weighting)) {
+
+    stop("\n-Data supplied to the weighting argument must be a two column data frame.",
+         call. = FALSE)
+
+  }
+
+
+
 #-------------------------------------------------------------
 #END OF CONDITIONS, NOW SUBSET LIBRARY
 #-------------------------------------------------------------
@@ -613,6 +681,12 @@ fps.powdRlib <- function(lib, smpl, harmonise, solver, obj, refs, std, force, st
 
     smpl <- harmonised$smpl
     lib <- harmonised$lib
+
+    #Predict weighting onto same scale
+    weighting <- stats::spline(x = weighting[[1]],
+                               y = weighting[[2]],
+                               method = "natural",
+                               xout = lib$tth)
 
   }
 
@@ -676,14 +750,26 @@ lib$xrd <- data.frame(lapply(lib$xrd,
 #Replace the library tth with that of the sample
 lib$tth <- smpl_tth
 
+if (!identical(weighting[[1]], lib$tth)) {
+
+#Predict weighting onto same scale
+weighting <- stats::spline(x = weighting[[1]],
+                           y = weighting[[2]],
+                           method = "natural",
+                           xout = lib$tth)
+
+}
+
+
+
 #### decrease 2TH scale to the range defined in the function call
-smpl <- smpl[which(smpl$tth >= tth_fps[1] & smpl$tth <= tth_fps[2]), ]
+#smpl <- smpl[which(smpl$tth >= tth_fps[1] & smpl$tth <= tth_fps[2]), ]
 
 #Subset the xrd dataframe too
-lib$xrd <- lib$xrd[which(lib$tth >= tth_fps[1] & lib$tth <= tth_fps[2]), , drop = FALSE]
+#lib$xrd <- lib$xrd[which(lib$tth >= tth_fps[1] & lib$tth <= tth_fps[2]), , drop = FALSE]
 
 #Replace the tth in the library with the shortened one
-lib$tth <- smpl[, 1]
+#lib$tth <- smpl[, 1]
 
 
 #-----------------------------------------------------------
@@ -691,15 +777,11 @@ lib$tth <- smpl[, 1]
 #-----------------------------------------------------------
 
 #The optimisation can fail if negative have creeped in during interpolation
-if(length(which(smpl[[2]] < 0) > 0) & obj == "Rwp") {
+if(min(smpl[[2]]) < 0 & obj == "Rwp") {
 
   cat("Negative values present in interpolated data. Switching objective
           function to R instead of Rwp to avoid errors.")
   obj <- "R"
-  #delete_negs <- which(smpl[[2]] < 0)
-  #smpl <- smpl[-delete_negs,]
-  #lib$tth <- lib$tth[-delete_negs]
-  #lib$xrd <- lib$xrd[-delete_negs, ]
 
 }
 
@@ -713,7 +795,9 @@ cat("\n-Optimising...")
 
 o <- stats::optim(par = x, .fullpat,
            method = solver, pure_patterns = lib$xrd,
-           sample_pattern = smpl[, 2], obj = obj)
+           sample_pattern = smpl[, 2], obj = obj,
+           tth = lib$tth, tth_fps = tth_fps,
+           weighting = weighting[[2]])
 
 x <- o$par
 
@@ -724,8 +808,9 @@ x <- o$par
 
 if (min(x) < 0) {
 
-remove_neg_out <- .remove_neg(x = x, lib = lib, smpl = smpl,
-                              solver = solver, obj = obj, force = force)
+remove_neg_out <- .remove_neg(x = x, lib = lib, smpl = smpl[[2]],
+                              solver = solver, obj = obj, force = force,
+                              tth_fps = tth_fps, weighting = weighting[[2]])
 
 x <- remove_neg_out[[1]]
 lib <- remove_neg_out[[2]]
@@ -735,7 +820,9 @@ lib <- remove_neg_out[[2]]
 } else {
 
   cat("\n-Applying non-negative least squares")
-  nnls_out <- .xrd_nnls(xrd.lib = lib, xrd.sample = smpl[, 2], force = force)
+  nnls_out <- .xrd_nnls(xrd.lib = lib, xrd.sample = smpl[[2]],
+                        force = force,
+                        tth_fps = tth_fps)
 
   lib$xrd <- nnls_out$xrd.lib
   x <- nnls_out$x
@@ -760,7 +847,8 @@ if(shift > 0 & length(x) > 1 & solver %in% c("Nelder-Mead", "BFGS", "CG")) {
   o <- stats::optim(par = x_s, .fullpat_shift_seq,
                     weightings = x,
                     method = solver, lib = lib,
-                    smpl = smpl, obj = obj)
+                    smpl = smpl, obj = obj,
+                    tth_fps = tth_fps)
 
   x_s <- o$par
 
@@ -779,6 +867,12 @@ if(shift > 0 & length(x) > 1 & solver %in% c("Nelder-Mead", "BFGS", "CG")) {
   lib <- shifted$lib
   smpl <- shifted$smpl
 
+  #Predict weighting onto same scale
+  weighting <- stats::spline(x = weighting[[1]],
+                             y = weighting[[2]],
+                             method = "natural",
+                             xout = lib$tth)
+
 #----------------------------------------------
 #Re-optimise after shifting
 #----------------------------------------------
@@ -786,21 +880,19 @@ if(shift > 0 & length(x) > 1 & solver %in% c("Nelder-Mead", "BFGS", "CG")) {
     cat("\n-Reoptimising after shifting data")
 
   #The optimisation can fail if negative have creeped in during interpolation
-  if(length(which(smpl[[2]] < 0) > 0) & obj == "Rwp") {
+  if(min(smpl[[2]]) < 0 & obj == "Rwp") {
 
     cat("Negative values present in interpolated data. Switching objective
          function to R instead of Rwp to avoid errors.")
     obj <- "R"
-    #delete_negs <- which(smpl[[2]] < 0)
-    #smpl <- smpl[-delete_negs,]
-    #lib$tth <- lib$tth[-delete_negs]
-    #lib$xrd <- lib$xrd[-delete_negs, ]
 
   }
 
     o <- stats::optim(par = x, .fullpat,
                       method = solver, pure_patterns = lib$xrd,
-                      sample_pattern = smpl[, 2], obj = obj)
+                      sample_pattern = smpl[, 2], obj = obj,
+                      tth = lib$tth, tth_fps = tth_fps,
+                      weighting = weighting[[2]])
 
     x <- o$par
 
@@ -811,8 +903,9 @@ if(shift > 0 & length(x) > 1 & solver %in% c("Nelder-Mead", "BFGS", "CG")) {
 
 if (min(x) < 0) {
 
-remove_neg_out <- .remove_neg(x = x, lib = lib, smpl = smpl,
-                              solver = solver, obj = obj, force = force)
+remove_neg_out <- .remove_neg(x = x, lib = lib, smpl = smpl[[2]],
+                              solver = solver, obj = obj, force = force,
+                              tth_fps = tth_fps, weighting = weighting[[2]])
 
 x <- remove_neg_out[[1]]
 lib <- remove_neg_out[[2]]
@@ -840,7 +933,7 @@ if (remove_trace > 0) {
 #--------------------------------------------------------
 
 #compute fitted pattern and residuals
-fitted_pattern <- apply(sweep(as.matrix(lib$xrd), 2, x, "*"), 1, sum)
+fitted_pattern <- apply(sweep(lib$xrd, 2, x, "*"), 1, sum)
 
 resid_x <- smpl[, 2] - fitted_pattern
 
@@ -874,29 +967,36 @@ if (is.na(std_conc) | identical(names(x), std)) {
 df <- min_concs[[1]]
 dfs <- min_concs[[2]]
 
-#Objective parameters for results
+#Get the index of the tth range in the function call
+tth_index <- which(lib$tth >= tth_fps[1] | lib$tth <= tth_fps[2])
+
+#Objective parameters for results (calculated with weighting vector)
 if (min(smpl[[2]]) < 0) {
 
   Rwp_fit <- NA
 
 } else {
 
-  Rwp_fit <- sqrt(sum((1/smpl[[2]]) * ((smpl[[2]] - fitted_pattern)^2)) / sum((1/smpl[[2]]) * (smpl[[2]]^2)))
+ Rwp_fit <- rwp(measured = smpl[tth_index, 2],
+                fitted = fitted_pattern[tth_index],
+                weighting = weighting[[2]])
 
 }
 
-R_fit <- sqrt(sum((smpl[[2]] - fitted_pattern)^2)/sum(smpl[[2]]^2))
+R_fit <- r(measured = smpl[tth_index, 2],
+           fitted = fitted_pattern[tth_index],
+           weighting = weighting[[2]])
 
-delta_fit <- sum(abs(smpl[[2]] - fitted_pattern))
+delta_fit <- delta(measured = smpl[tth_index, 2],
+                   fitted = fitted_pattern[tth_index],
+                   weighting = weighting[[2]])
 
 #Extract the xrd data
 xrd <- data.frame(lib$xrd,
                   check.names = FALSE)
 
 #Scale them by the optimised weightings
-for (i in 1:ncol(xrd)) {
-  xrd[,i] <- xrd[,i] * x[i]
-}
+xrd <- sweep(xrd, 2, x, "*")
 
 #If only 1 pattern is used in the fit, then rename it
 if (ncol(xrd) == 1) {
@@ -917,7 +1017,8 @@ inputs <- list("harmonise" = harmonise,
                "manual_align" = manual_align,
                "tth_fps" = tth_fps,
                "shift" = shift,
-               "remove_trace" = remove_trace)
+               "remove_trace" = remove_trace,
+               "weighting" = weighting[[2]])
 
 
 #Define a list that becomes the function output